Refactor background components to dedicated directory structure
- Moved beams-background.tsx to backgrounds/beams directory
- Moved wavy-background.tsx to backgrounds/wavy directory
- Created consistent directory structure with index.ts and README.md files
- Added backward compatibility re-exports with @deprecated annotations
- Added main backgrounds/index.ts file and README.md
- Updated COMPONENT-MOVEMENT-PLAN.md to track changes
