+++ src/components/ui/backgrounds/beams/index.ts
@@ -1,0 +1,1 @@
+export * from "./beams-background";
