+++ src/components/ui/backgrounds/gradient/index.ts
@@ -1,0 +1,1 @@
+export * from "./background-gradient";
