+++ src/components/ui/backgrounds/index.ts
@@ -1,0 +1,4 @@
+export * from "./beams";
+export * from "./gradient";
+export * from "./warp";
+export * from "./wavy";
