+++ src/components/ui/backgrounds/wavy/index.ts
@@ -1,0 +1,1 @@
+export * from "./wavy-background";
